# Optimize string comparisons