See https://github.com/libgit2/libgit2/issues/4230#issuecomment-471710359 for context. This is diff optimization PR 1/3. The smallest of the three both in terms of code delta and impact.

Three hand-rolled string comparison functions -- git__prefixcmp, git__strcmp and git__strncmp -- are showing up on the CPU profile when scanning working directory of chromium repository on Linux. This PR optimizes git__prefixcmp and replaces the other two functions with standard equivalents (unless `GIT_WIN32` is defined).

Test:

```zsh
cd ~/libgit2
mkdir build
cd build
cmake \
  -DUSE_ICONV=OFF           \
  -DBUILD_CLAR=ON           \
  -DUSE_SSH=OFF             \
  -DUSE_HTTPS=OFF           \
  -DTHREADSAFE=OFF          \
  -DBUILD_SHARED_LIBS=OFF   \
  -DUSE_EXT_HTTP_PARSER=OFF \
  -DUSE_BUNDLED_ZLIB=ON     \
  -DBUILD_EXAMPLES=ON       \
  -DCMAKE_BUILD_TYPE=Release ..
make -j 20
./libgit2_clar
```

Benchmark:

```zsh
git clone https://github.com/chromium/chromium.git ~/chromium
cd ~/chromium
~/libgit2/build/examples/lg2 status >/dev/null &&
  time (repeat 100; do ~/libgit2/build/examples/lg2 status >/dev/null; done)
```

  * Baseline: 112.11s user 62.75s system 99% cpu 2:55.15 total
  * With this PR: 104.80s user 62.93s system 99% cpu 2:47.96 total

A modest 4% improvement. Working directory traversal increases more than that. The benchmark spends a lot of time doing other things, so the apparent impact of the PR is diluted.

Note:

  * Built and tested only on Ubuntu 18.04 with gcc 8.2. I didn't even attempt to compile on Windows.
  * Benchmarked only on chromium repository on a single machine running Ubuntu 18.04.