@romkatv 2019-03-19
> I guess you didn't take a lot at the generated assembly, right?

Right, I haven't look at the assembly. This PR was produced with the following methodology.

1. Run `time lg2 status` on chromium repo.
2. Generate CPU profile for `lg2 status` in chromium repo. I used perf + perf_data_converter + pprof for this.
3. Take one of the hand-rolled string comparison functions from the top CPU consumers and replace it with a standard function.
4. Run `time lg2 status` on chromium repo. If it's faster than before, keep the change. Otherwise revert it.
5. Repeat.